Tidy Up and move SelectableGroup
