--- docs/src/pages/lab/selectable-group/selectable-group.md
@@ -1,17 +1,0 @@
----
-title: Selectable Group React component
-components: SelectableGroup
----
-
-# Selectable Group
-
-<p class="description">A Selectable Group manages all the selection state</p>
-
-#### Select Many
-
-{{"demo": "pages/lab/selectable-group/Checkboxes.js"}}
-
-#### Select One
-
-{{"demo": "pages/lab/selectable-group/RadioButtons.js"}}
-
