+++ packages/material-ui/src/internal/SelectableGroup/useSelectedState.js
@@ -1,6 +1,9 @@
 import React from 'react';
 import SelectableGroupContext from './SelectableGroupContext';
 
+/**
+ * @ignore - internal component.
+ */
 function useSelectedState() {
   return React.useContext(SelectableGroupContext);
 }
