+++ pages/_app.js
@@ -284,9 +284,6 @@ const pages = [
         title: 'About The Lab',
       },
       {
-        pathname: '/lab/selectable-group',
-      },
-      {
         pathname: '/lab/slider',
       },
       {
