--- pages/lab/api/selectable-group.js
@@ -1,11 +1,0 @@
-import 'docs/src/modules/components/bootstrap';
-// --- Post bootstrap -----
-import React from 'react';
-import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
-import markdown from './selectable-group.md';
-
-function Page() {
-  return <MarkdownDocs markdown={markdown} />;
-}
-
-export default Page;
