--- pages/lab/api/selectable-group.md
@@ -1,32 +1,0 @@
----
-filename: /packages/material-ui-lab/src/SelectableGroup/SelectableGroup.js
----
-
-<!--- This documentation is automatically generated, do not try to edit it. -->
-
-# SelectableGroup API
-
-<p class="description">The API documentation of the SelectableGroup React component. Learn more about the properties and the CSS customization points.</p>
-
-```js
-import SelectableGroup from '@material-ui/lab/SelectableGroup';
-```
-
-
-
-## Props
-
-| Name | Type | Default | Description |
-|:-----|:-----|:--------|:------------|
-| <span class="prop-name">additional</span> | <span class="prop-type">object</span> |   | Additional props to pass down. |
-| <span class="prop-name">children</span> | <span class="prop-type">node</span> |   | The contents of the selectable group. |
-| <span class="prop-name">exclusive</span> | <span class="prop-type">bool</span> | <span class="prop-default">false</span> | Only allows one value to be selected. |
-| <span class="prop-name">onChange</span> | <span class="prop-type">func</span> |   | Functioned called when value changes. |
-| <span class="prop-name">value</span> | <span class="prop-type">any</span> |   | The currently selected value within the group or an array of selected values when `exclusive` is false. |
-
-Any other properties supplied will be spread to the root element (native element).
-
-## Demos
-
-- [Selectable Group](/lab/selectable-group/)
-
