--- pages/lab/selectable-group.js
@@ -1,18 +1,0 @@
-import 'docs/src/modules/components/bootstrap';
-// --- Post bootstrap -----
-import React from 'react';
-import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
-
-const req = require.context('docs/src/pages/lab/selectable-group', false, /\.md|\.js$/);
-const reqSource = require.context(
-  '!raw-loader!../../docs/src/pages/lab/selectable-group',
-  false,
-  /\.js$/,
-);
-const reqPrefix = 'pages/lab/selectable-group';
-
-function Page() {
-  return <MarkdownDocs req={req} reqSource={reqSource} reqPrefix={reqPrefix} />;
-}
-
-export default Page;
